requirements.txt
asgiref==3.7.2
async-generator==1.10
atomicwrites==1.4.1
attrs==23.1.0
backports.zoneinfo==0.2.1
black==23.3.0
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
docutils==0.19
exceptiongroup==1.1.1
factory-boy==3.2.1
Faker==18.10.1
filelock==3.12.2
h11==0.14.0
identify==2.5.24
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
mypy-extensions==1.0.0
nodeenv==1.8.0
outcome==1.2.0
packaging==23.1
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.6.0
pluggy==1.0.0
pre-commit==3.3.3
py==1.11.0
pycparser==2.21
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-factoryboy==2.5.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-selenium==4.0.0
pytest-variables==2.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.3
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
wsproto==1.2.0
